Change % to %% when using ${DEFAULT-VALUE} in option description
Percent signs shall be escaped in the description text as it is going
to be passed to String.format. If the default value contains (single) %
signs then using the built-in DEFAULT-VALUE variable would lead to a
warning message. This is avoided by escaping any % signs in the
DEFAULT-VALUE variable when used in the description.