Adding picocli-core-module
Addong a JMPS 9 compliant (hopefully) subproject. This subproject needs the root project to work as source files are copied to this subproject before compilation.

It assumes that tests are run from the root project and not against this subproject, the objective being only to release a compliant Java 9 JPMS module.