[#667] Embed GraalVM configuration in picocli jar
Decided not to do this, but committing this for reference.
There is a circular dependency that makes it difficult to update the
main picocli jar with a file generated by a subproject.

One solution is to maintain a resource file in the main project in
/src/main/resources/META-INF/native-image/info.picocli/picocli/reflect-config.json,
and update this file every time the subproject is built.
Next time we commit the main project, this resource file would also be committed.