Add command() to ArgSpec
There seems to be no public way to get from an ArgSpec to the command
model it belongs to. Expose the existing field with a getter method.

My use case for this is so I can write a generic IDefaultValueProvider
that is sensitive to which subcommand it is getting a default for. This
makes it simpler to implement Bazel-style config files that provide a
uniform way to override defaults for all command-line options.