Remove flaky findOriginal() method and add .root() to ArgSpec
The best-effort matching search to add the root element was changed to
simply wire up the root ArgSpec to each of the inherited arguments.

Arguments which are not inheritable will have a null root, and
inheritable arguments (options or positional parameters) will have a
root() method that returns the root argument, or the argument itself if
it is already at the root.

Unit tests have also been added to check that this behavior is
maintained.