Added test class for issue $1409. Classname ArgGroupDefaultValueTest
Description of issue:
I understand that by default defaultValue in ArgGroup are not applied.

The behavior is explained at : #876 (comment)
(Note: The default behavior seems surprising to me at first sight but maybe it's justified)

I faced a situation where it seems that this behavior is not respected (but maybe I missed something)
Sorry for the not so short example, I was not able to make it smaller.

The goal is to have this kind of options : X AND ((1A AND 1B) OU (2A OU 2B))
I set default value for X, 2A and 2B.