1615: Fix bug in getCJKAdjustedLength() that double counts supplementary code points
getCJKAdjustedLength calculates the column width of text strings.

But it operates on chars which causes it to double count supplementary code
points. Supplimentary code points have a high and a low code unit(char).

This commit reworks this logic to be based on code points. The code assembles
the code points by hand as String.codePoints() was only added in 1.8 and
picocli has a source level of 1.5